+++ src/index.ts
@@ -6,3 +6,6 @@ export type { ChipSelectorProps, ChipOption } from './ChipSelector/ChipSelector'
 
 export { FormSwitch } from './FormSwitch/FormSwitch';
 export type { FormSwitchProps } from './FormSwitch/FormSwitch';
+
+export { FormCheckbox } from './FormCheckbox/FormCheckbox';
+export type { FormCheckboxProps } from './FormCheckbox/FormCheckbox';